\name{print.teigen}
\alias{print.teigen}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
print.teigen: Print Function for tEIGEN Objects
}
\description{
Outputs short, concise information on the teigen object: BIC value, best model, and best group. Same info for ICL is output if it disagrees with the BIC value.
}
\usage{
\method{print}{teigen}(x, \dots)
}

\arguments{
  \item{x}{
An object of class \code{\link{teigen}}
}
\item{\dots}{
Options to be passed to \code{print}.
}
}


\author{
%%  ~~who you are~~
Jaymeson R. Wickins, Nicholas M. Boers, Jeffrey L. Andrews
}


%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{teigen}}
}

