\name{summary.teigen}
\alias{summary.teigen}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
summary.teigen: Summary Function for tEIGEN Objects
}
\description{
Gives summary information.
}
\usage{
\method{summary}{teigen}(object, \dots)
}

\arguments{
  \item{object}{
An object of class \code{\link{teigen}}
}
\item{\dots}{
Options to be passed to \code{summary}.
}
}
\author{
%%  ~~who you are~~
Jeffrey L. Andrews
}


%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{teigen}}
}
