\name{teigen-package}
\alias{teigen-package}
\alias{teigenpackage}
\alias{tEIGEN}
\docType{package}
\title{
teigen: Model-based clustering and classification with the multivariate t-distribution
}
\description{
Fits mixtures of multivariate t-distributions (with eigen-decomposed covariance structure) via the multi-cycle ECM algorithm under a clustering or classification paradigm.
}
\details{
\tabular{ll}{
Package: \tab teigen\cr
Type: \tab Package\cr
Version: \tab 2.0.7\cr
Date: \tab 2014-08-19\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Jeffrey L. Andrews, Paul D. McNicholas

Maintained by: Jeffrey L. Andrews <jeffrey.andrews@macewan.ca>
}
\references{
Andrews JL and McNicholas PD (2012). ``Model-based clustering, classification, and discriminant analysis with the multivariate \emph{t}-distribution: The \emph{t}EIGEN family'' \emph{Statistics and Computing} 22(5), 1021--1029. 

Andrews JL, McNicholas PD, and Subedi S (2011) ``Model-based classification via mixtures of multivariate t-distributions'' \emph{Computational Statistics and Data Analysis} 55, 520--529.
}
\keyword{ package }
\seealso{
\code{\link{teigen}} for main function
}
