\name{print.teigen}
\alias{print.teigen}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
print.teigen: Print Function for tEIGEN Objects
}
\description{
Outputs summary information.
}
\usage{
\method{print}{teigen}(x, \dots)
}

\arguments{
  \item{x}{
An object of class \code{\link{teigen}}
}
\item{\dots}{
Options to be passed to \code{print}.
}
}


\author{
%%  ~~who you are~~
Jeffrey L. Andrews
}


%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{teigen}}
}

