\name{teigen-package}
\alias{teigen-package}
\alias{teigenpackage}
\alias{tEIGEN}
\docType{package}
\title{
teigen: Model-based clustering and classification with the multivariate t-distribution
}
\description{
Fits multivariate t-distribution mixture models (with eigen-decomposed covariance structure) to the given data underneath a clustering or classification paradigm.
}
\details{
\tabular{ll}{
Package: \tab teigen\cr
Type: \tab Package\cr
Version: \tab 2.0.1\cr
Date: \tab 2013-10-03\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Jeffrey L. Andrews, Paul D. McNicholas

Maintained by: Jeffrey L. Andrews <jeffrey.andrews@macewan.ca>
}
\references{
Andrews JL and McNicholas PD. ``Model-based clustering, classification, and discriminant analysis with the multivariate \emph{t}-distribution: The tEIGEN family'' \emph{Statistics and Computing}. To appear: http://dx.doi.org/10.1007/s11222-011-9272-x

Andrews JL, McNicholas PD, and Subedi S (2011) ``Model-based classification via mixtures of multivariate t-distributions'' \emph{Computational Statistics and Data Analysis} 55, 520--529.
}
\keyword{ package }
\seealso{
\code{\link{teigen}} for main function
}
