\name{teigen-package}
\alias{teigen-package}
\alias{teigenpackage}
\alias{tEIGEN}
\docType{package}
\title{
teigen: Model-Based Clustering and Classification with the Multivariate t Distribution
}
\description{
Fits mixtures of multivariate t-distributions (with eigen-decomposed covariance structure) via the expectation conditional-maximization algorithm under a clustering or classification paradigm.
}
\details{
\tabular{ll}{
Package: \tab teigen\cr
Type: \tab Package\cr
Version: \tab 2.1.0\cr
Date: \tab 2015-11-19\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Jeffrey L. Andrews, Jaymeson R. Wickins, Nicholas M. Boers, Paul D. McNicholas

Maintained by: Jeffrey L. Andrews <jeff.andrews@ubc.ca>
}
\references{
Andrews JL and McNicholas PD (2012). ``Model-based clustering, classification, and discriminant analysis with the multivariate \emph{t}-distribution: The \emph{t}EIGEN family'' \emph{Statistics and Computing} 22(5), 1021--1029. 

Andrews JL, McNicholas PD, and Subedi S (2011) ``Model-based classification via mixtures of multivariate t-distributions'' \emph{Computational Statistics and Data Analysis} 55, 520--529.
}
\keyword{ package }
\seealso{
\code{\link{teigen}} for main function
}
